Move guardedOpenModal() into the modal component file
This logic is strongly connected to the modal component.
